Overhauled course parser to use common spaces (broken)
Instead of splitting by space and using some complex logic to
detemine what each element is, this approach looks across all
results of a course for common spaces and uses these to define
column boundaries.

However, this fails in practice because there are a decent
number of rows that skip spaces due to long entries. These invalid
rows prevent the entire course being parsed. A solution is being
worked on, but this is being committed because it is cool imo.